Restructure state to make it more extendable
"condition" is the domain name for one clause or group in the
QueryRepresentation.  The state for a condition, a ConditionRow, will in
the future both contain the data needed for the query itself and some
metadata like the label or the error.

The Query representation will be generated by the getter from the
RootState. Eventually, we will need to be a (lossy) two-way transformation
between the Store's RootState and QueryRepresentation.
From that QueryRepresentation we willhave then a one-way transformation into
SPARQL and another two-way transformation into JSON(?) for encoding it
in the URL.

Getters should be named consistently when used with property-style access.