Push validation into the store
Handling the validation in the components is becoming too complex with
multiple rows. Pushing it into the store separates the business logic
from the display logic.

This also fixes some issues with the a11y test, it should now log less
messages in the console.

It incidentally also resolves T269420 as I had to move that code anyway
and so it was easy to just drop the line that was causing the problems.

This is done in preparation for T269420