Wire up AddCondition button
This adds a "conditionId" field to the "ConditionRow". This is done
because we need a stable identifier as the key in v-for.

However, we cannot use this id as the key of an object, because such an
object would not have an inherent order and we would have to maintain an
extra array with all the keys to define the order. This complexity does
not seem worth it.

Bug: T269136