Use grid layout in QueryCondition
This adds the placeholders for the planned toggle buttons and references
Dropdown in order to be able to take them into account for the layout.
A dark border was added between the buttons to prevent them from looking
like one single disabled button.

The magic number for the $clearWidth comes from the magic number for the
intro text. Not ideal, but works for now. Maybe this should be replace
by some em value?

This doesn't work too nicely for very small screens, but it is still
usable.

Bug: T269470