fix: debounce searching for options on typing
The debounce in the repository wasn't working as it was recreated on
every call.

Unfortunately, lodash.debounce does not properly work with methods that
actually return something as it seemingly always returns undefined on
first invocation.

Therefore, the debounce is moved to the EntityLookup, and the update to
the state happens inside the debounced method.

A delay of 150ms was chosen, because 100ms doesn't seem long enough to
debounce many events for a average type like me and 300ms is very
noticeable. 150ms seemed like a good value in between.

Bug: T273868